refactor: make stonks order resolver form available for everyone
