+++ modules/stonks/ui/StonksGridWrapper/index.ts
@@ -1,0 +1,1 @@
+export { StonksGridWrapper } from './StonksGridWrapper'
